requirements.txt
asgiref==3.8.1 ; python_version == "3.12"
django-cors-headers==4.7.0 ; python_version == "3.12"
django==5.2.3 ; python_version == "3.12"
djangorestframework-simplejwt==5.5.0 ; python_version == "3.12"
djangorestframework==3.16.0 ; python_version == "3.12"
pyjwt==2.9.0 ; python_version == "3.12"
sqlparse==0.5.3 ; python_version == "3.12"
tzdata==2025.2 ; python_version == "3.12" and sys_platform == "win32"
